+++ music_queue/extensions.py
@@ -14,7 +14,7 @@ if os.getenv("LLM_SOURCE") == "huggingface":
         lrc_pipeline = transformers.pipeline(
             "text-generation",
             model=lrc_model_id,
-            model_kwargs={"torch_dtype": torch.bfloat16},
+            model_kwargs={"torch_dtype": torch.float32},
             device_map="auto",
         )
         print(f"LRC Model {lrc_model_id} loaded successfully!")
@@ -25,14 +25,14 @@ migrate = Migrate()
 logging.basicConfig()
 logging.getLogger("sqlalchemy.engine").setLevel(logging.INFO)
 
+
 def register_extensions(app, worker=False):
-    
+
     db.init_app(app)
     migrate.init_app(app, db)
 
     logging.info(f"Pipeline source -> {os.getenv("LLM_SOURCE")}")
-    
-    
+
     from music_shared.models import (
         Music,
         Prompt,
